+++ src/main/kotlin/io/github/jamalam/database/DefaultValues.kt
@@ -35,7 +35,8 @@ fun KClass<ServerConfig>.getDefault(id: Snowflake): ServerConfig {
         ServerModerationConfig::class.getDefault(),
         ServerTagsConfig::class.getDefault(),
         ServerNotificationsConfig::class.getDefault(),
-        ServerFilePasteConfig::class.getDefault()
+        ServerFilePasteConfig::class.getDefault(),
+        ServerPhishingConfig::class.getDefault(),
     )
 }
 
@@ -82,5 +83,11 @@ fun KClass<ServerFilePasteConfig>.getDefault(): ServerFilePasteConfig {
     return ServerFilePasteConfig(
         enabled = false,
         hastebinUrl = "https://www.toptal.com/developers/hastebin/"
+    )
+}
+
+fun KClass<ServerPhishingConfig>.getDefault(): ServerPhishingConfig {
+    return ServerPhishingConfig(
+        enabled = false
     )
 }